Add another 8-bit field in vtkDICOMValue for overflow.
This is specifically for values that use delimiters instead of fixed
lengths, such as OB or UN VRs that store encapsulated data.  By adding
an 8-bit overflow byte for the 32-bit NumberOfValues field, it becomes
possible to store encapsulated data with sizes up to 2^40 bytes (1TB).
Values that use fixed lengths are still limited to 2^32-2 bytes (4GB)
as per the standard.

+++ vtkDICOMSequence.cxx
@@ -22,7 +22,7 @@ const vtkDICOMValue vtkDICOMSequence::InvalidValue;
 
 //----------------------------------------------------------------------------
 const vtkDICOMValue &vtkDICOMSequence::GetAttributeValue(
-    unsigned int i, vtkDICOMTag tag) const
+    size_t i, vtkDICOMTag tag) const
 {
   const vtkDICOMItem *ptr = this->V.GetSequenceData();
   if (ptr != 0 && i < this->V.GetNumberOfValues())
@@ -34,7 +34,7 @@ const vtkDICOMValue &vtkDICOMSequence::GetAttributeValue(
 
 //----------------------------------------------------------------------------
 const vtkDICOMValue &vtkDICOMSequence::GetAttributeValue(
-    unsigned int i, const vtkDICOMTagPath &tagpath) const
+    size_t i, const vtkDICOMTagPath &tagpath) const
 {
   const vtkDICOMItem *ptr = this->V.GetSequenceData();
   if (ptr != 0 && i < this->V.GetNumberOfValues())
@@ -45,7 +45,7 @@ const vtkDICOMValue &vtkDICOMSequence::GetAttributeValue(
 }
 
 //----------------------------------------------------------------------------
-const vtkDICOMItem &vtkDICOMSequence::GetItem(unsigned int i) const
+const vtkDICOMItem &vtkDICOMSequence::GetItem(size_t i) const
 {
   const vtkDICOMItem *ptr = this->V.GetSequenceData();
   assert(ptr != 0 && i < this->V.GetNumberOfValues());
